// frontend/src/types/update-lead.ts
export interface UpdateLeadRequest {

  id: string;

  //////////////////////////////////////////////////
  // BASIC
  //////////////////////////////////////////////////

  name: string;

  company: string;

  email: string | null;

  phone: string | null;

  //////////////////////////////////////////////////
  // ADDRESS
  //////////////////////////////////////////////////

  address: string | null;

  district: string | null;

  city: string | null;

  province: string | null;

  postalCode: string | null;

  country: string | null;

  //////////////////////////////////////////////////
  // RELATION
  //////////////////////////////////////////////////

  sourceId: string;

  assignedTo: string;

  //////////////////////////////////////////////////
  // PIPELINE
  //////////////////////////////////////////////////

  status: string;

}